fix: Missing global variable of external packages.
When you run `yarn build`, there will be a warning like this:
```
No name was provided for external module 'react' in output.globals – guessing 'React'
```

Cause:
UMD build is a universal build, which means it act as amd, cjs and iife build.
As iife it needs to define global variable to replace external module that being used in runtime.
Which in this case is `react`.

Solution:
Add in vite.config.js, in build.rollupUptions block:
```js
output: {
  globals: {
    'react': 'React'
  },
}
```js